removed netcdf_utils
#244
