--- tests/test_netcdf_utils.py
@@ -1,47 +1,0 @@
-import logging
-
-import numpy as np
-import pytest
-
-from scripts import netcdf_utils
-
-# ----------------------------------------------------------------------------------------------------------------------
-# disable logging messages
-logging.disable(logging.CRITICAL)
-
-
-# Tests for `scripts.netcdf_utils.py`
-
-
-# ---------------------------------------------------------------------------------------
-def test_find_netcdf_datatype():
-    """
-    Test for the netcdf_utils.find_netcdf_datatype() function
-    """
-
-    data_objs = (
-        ["f8", float(1.0)],
-        ["f4", np.float32(1.0)],
-        ["f8", np.float64(1.0)],
-        ["f8", np.double(1.0)],
-        ["f8", np.NaN],
-        ["i2", np.int16(1.0)],
-        ["i4", np.int32(1.0)],
-        ["i4", int(1.0)],
-    )
-    for [data_type, data_object] in data_objs:
-
-        if data_type != netcdf_utils.find_netcdf_datatype(data_object):
-            raise AssertionError(
-                "Failed to determine expected data type for {0}, expected {1}".format(
-                    data_object, data_type
-                )
-            )
-
-    # provide nonsense arguments, should raise error in each case
-    with pytest.raises(ValueError):
-        netcdf_utils.find_netcdf_datatype(data_object={"abc": 123})
-        netcdf_utils.find_netcdf_datatype("hokey pokey")
-        netcdf_utils.find_netcdf_datatype(
-            ["this is a list of heterogeneous items", 14, {43: 56}]
-        )
